+++ actionpack/lib/action_controller/test_case.rb
@@ -360,7 +360,6 @@ module ActionController
       extend ActiveSupport::Concern
       include ActionDispatch::TestProcess
       include ActiveSupport::Testing::ConstantLookup
-      include Rails::Dom::Testing::Assertions
 
       attr_reader :response, :request
 
@@ -665,10 +664,6 @@ module ActionController
           env.delete "CONTENT_LENGTH"
           env.delete "RAW_POST_DATA"
           env
-        end
-
-        def document_root_element
-          html_document.root
         end
 
         def check_required_ivars