+++ actionpack/lib/action_dispatch/testing/integration.rb
@@ -681,10 +681,6 @@ module ActionDispatch
       def app
         super || self.class.app
       end
-
-      def document_root_element
-        html_document.root
-      end
     end
 
     include Behavior