+++ actionmailer/CHANGELOG.md
@@ -1,3 +1,10 @@
+*   Introduce `config.action_mailer.html_assertions`
+
+    Adds support for testing with `Capybara::Minitest::Assertions` when set to `:capybara`.
+    Defaults to `Rails::Dom::Testing::Assertions` with `:rails_dom_testing`.
+
+    *Sean Doyle*
+
 *   Remove deprecated params via `:args` for `assert_enqueued_email_with`.
 
     *Rafael Mendonça França*