+++ actionpack/CHANGELOG.md
@@ -1,5 +1,7 @@
-*   Introduce `ActionDispatch::Assertions::CapybaraAssertions` to support
-    `Capybara` assertions in `ActionDispatch::IntegrationTest`
+*   Introduce `config.action_controller.html_assertions` and `config.action_dispatch.html_assertions`
+
+    Adds support for testing with `Capybara::Minitest::Assertions` when set to `:capybara`.
+    Defaults to `Rails::Dom::Testing::Assertions` with `:rails_dom_testing`.
 
     *Sean Doyle*